code: ensure modals work on screen sizes that are over the height of the viewport
However, the sucky tradeoff here is that modals aren't vertically centered
on the screen. I took a look at how bootstrap does it and they don't
vertically center modals probably for this exact reason. Take a look:
https://getbootstrap.com/docs/4.0/components/modal/

Ideally, modals could be centered. But I'm not sure if there's a way to
do this without resorting to javascript.