card: add key for table off of index (in additon to other things) to work around a key uniqueness issue
To summarize, the key for an item cannot be the item since you could
have two cells that contain the same value in the table. I think a
combination of the row id, the cell value, and the index should be
unique while still ensuring that react will rerender a row or column
when it changes.