code: hourly breakdown accepts all data and filters it
Previously, the hourly breakdown report accepted only the data it needed
to show, and then used the value provided in `dataStartTime` to offset
the data by the hour it started on. This isn't ideal when expanding this
report though, because then the data would have to be refetched in
order to display more of it, because the data that wasn't "in view" in
the smaller version of the report was already thrown away.

So instead, the report now accepts all data, and the data is filtered
when on a report that shows less data by specifying `dataStartTime`
and `dataStartEnd` props. When on a larger report, these values can be
changed without the data itself having to be refetched.