code: forward ref passed to InputBox to input element
This is so that I can autofocus an input when it renders. In addition, I
upgraded to the latest storybook version, as I was running into this
issue, and it was fixed in storybook 5.0.6:
https://github.com/storybooks/storybook/issues/5974#issuecomment-478434674

An alternative to this approah could be to add a "focus" method to the
component instance. Because the ref prop can be used to get a component
instance in addition, this may be abstracted a bit nicer. However, I
tend to think this approach would be more confusing than helpful and not
all that intutive to someone using the component for the first time.