code: fix issue with tag input z index
By default before, the wrapper div around the control always had a very
large z-index. This meant that it would often be above other controls
when they were opened, which would not be expected. To get around this,
only set the z-index to be large whe nthe control has focus / is open.